+++ runtime/src/meta-stores/InlineMetadataStore.ts
@@ -39,17 +39,8 @@ export class InlineMetadataStore extends MetadataStoreBase
 
 	getLazy(id: number): () => (Type | undefined)
 	{
-		return function () {
+		return function lazyType() {
 			return InlineMetadataStore.get().get(id) ?? undefined;
 		};
-	}
-
-	set(id: number, description: any): Type
-	{
-		const type = this.wrap(description);
-
-		this._store[id] = type;
-
-		return type;
 	}
 }