+++ runtime/src/meta-stores/WindowMetadataStore.ts
@@ -56,17 +56,8 @@ export class WindowMetadataStore extends MetadataStoreBase
 
 	getLazy(id: number): () => (Type | undefined)
 	{
-		return function () {
+		return function lazyType() {
 			return window[REFLECT_STORE_SYMBOL].get(id) ?? undefined;
 		};
-	}
-
-	set(id: number, description: any): Type
-	{
-		const type = this.wrap(description);
-
-		this._store[id] = type;
-
-		return type;
 	}
 }