+++ tests/src/12-generic-property-decorator.ts
@@ -15,9 +15,33 @@ test("Property decorator reflects generic type", () => {
 		expect(type.name).toBe("Something");
 	}
 
+	function reference<TType>(name: string, description: string)
+	{
+		expect(name).toBe("Louis Litt");
+		expect(description).toBe("description");
+
+		const type = getType<TType>();
+		expect(type instanceof Type).toBe(true);
+		expect(type).not.toBe(Type.Unknown);
+		expect(type.name).toBe("Something");
+
+		return (target: Object, key: string | symbol) => {
+			
+		};
+	}
+
+	/**
+	 * @reflect
+	 */
+	function forward<TType>(name: string)
+	{
+		return reference<TType>(name, "description");
+	}
+
 	class Something
 	{
 		@propertyDecorator
+		@forward("Louis Litt")
 		property: string = "";
 
 		method(): string